Exposed Bootloader version available when booting a device in bootloader state
