+++ src/device/DeviceBootloader.cpp
@@ -1333,71 +1333,6 @@ std::vector<std::uint8_t> DeviceBootloader::getEmbeddedBootloaderBinary(Type typ
     return Resources::getInstance().getBootloaderFirmware(type);
 }
 
-DeviceBootloader::Version::Version(const std::string& v) : versionMajor(0), versionMinor(0), versionPatch(0), buildInfo{""} {
-    // Parse string
-    char buffer[256]{0};
-    if(std::sscanf(v.c_str(), "%u.%u.%u+%255s", &versionMajor, &versionMinor, &versionPatch, buffer) != 4) {
-        if(std::sscanf(v.c_str(), "%u.%u.%u", &versionMajor, &versionMinor, &versionPatch) != 3) {
-            throw std::runtime_error("Cannot parse version: " + v);
-        }
-    } else {
-        buildInfo = std::string{buffer};
-    }
-}
-
-DeviceBootloader::Version::Version(unsigned vmajor, unsigned vminor, unsigned vpatch)
-    : versionMajor(vmajor), versionMinor(vminor), versionPatch(vpatch), buildInfo{""} {}
-
-DeviceBootloader::Version::Version(unsigned vmajor, unsigned vminor, unsigned vpatch, std::string buildInfo)
-    : versionMajor(vmajor), versionMinor(vminor), versionPatch(vpatch), buildInfo(buildInfo) {}
-
-bool DeviceBootloader::Version::operator==(const Version& other) const {
-    if(versionMajor == other.versionMajor && versionMinor == other.versionMinor && versionPatch == other.versionPatch && buildInfo == other.buildInfo) {
-        return true;
-    }
-    return false;
-}
-
-bool DeviceBootloader::Version::operator<(const Version& other) const {
-    if(versionMajor < other.versionMajor) {
-        return true;
-    } else if(versionMajor == other.versionMajor) {
-        if(versionMinor < other.versionMinor) {
-            return true;
-        } else if(versionMinor == other.versionMinor) {
-            if(versionPatch < other.versionPatch) {
-                return true;
-            } else if(versionPatch == other.versionPatch) {
-                if(!buildInfo.empty() && other.buildInfo.empty()) {
-                    return true;
-                }
-            }
-        }
-    }
-    return false;
-}
-
-std::string DeviceBootloader::Version::toString() const {
-    std::string version = std::to_string(versionMajor) + "." + std::to_string(versionMinor) + "." + std::to_string(versionPatch);
-    if(!buildInfo.empty()) {
-        version += "+" + buildInfo;
-    }
-    return version;
-}
-
-std::string DeviceBootloader::Version::toStringSemver() const {
-    std::string version = std::to_string(versionMajor) + "." + std::to_string(versionMinor) + "." + std::to_string(versionPatch);
-    return version;
-}
-
-std::string DeviceBootloader::Version::getBuildInfo() const {
-    return buildInfo;
-}
-
-DeviceBootloader::Version DeviceBootloader::Version::getSemver() const {
-    return Version(versionMajor, versionMinor, versionPatch);
-}
-
 template <typename T>
 bool DeviceBootloader::sendRequest(const T& request) {
     if(stream == nullptr) return false;
